# Get only those block items which have their path to root