[SUST-20 Improved](https://openedx.atlassian.net/browse/SUST-20)

This PR is redo of edx/edx-platform#10994 which was reverted due to its implication on performance. We can use `get_items` with `include_orphans ` set to `False`, this will help us getting only those items which are not orphans. Whenever we are not concerned about getting orphans, just don't pass `include_orphans` kwarg while calling `get_items`.


**Performance Comparison:**

_With reverted PR , `get_items` was taking 386 milliseconds from which total 4 calls to [`load_tagged_classes`](https://github.com/edx/edx-platform/commit/0ae51c2142a880e8c014859c067bb60a08767bdf#diff-01698fa4b2435cbab4063c626df19d9bR1126) were consuming 343 milliseconds as shown in profile_
<img width="680" alt="reverted_get_items_tracelog_1" src="https://cloud.githubusercontent.com/assets/7848408/12066052/229663a8-b003-11e5-9f6b-b538245b0e50.png">
<img width="789" alt="reverted_get_items_tracelog_2" src="https://cloud.githubusercontent.com/assets/7848408/12066062/346efd6a-b003-11e5-9ba4-b2bebb2c0195.png">

_With this PR , `get_items` is taking only 44 milliseconds when called with or without `include_orphans=False ` as shown in profiles below_

<img width="731" alt="improved_get_items" src="https://cloud.githubusercontent.com/assets/7848408/12066127/fee0eda6-b003-11e5-91cc-dc2417b64413.png">
<img width="724" alt="with include_orphans false" src="https://cloud.githubusercontent.com/assets/7848408/12066132/0d8ae8a2-b004-11e5-9b45-3a39bd140716.png">