@adampalay 2015-12-30
@Qubad786 , can you please post some indication of the performance improvement between the two PRs on the ticket?