@Qubad786 2015-12-31
@adampalay I have updated PR with performance comparison.