@adampalay 2016-01-04
@Qubad786 , one minor point, otherwise this looks good. Thanks for the performance comparison, that's awesome