@adampalay 2016-01-04
So to be clear, this PR makes `get_items` 0.044 - 0.034 = 0.01 seconds slower?