@Qubad786 2016-01-04
In reverted PR , `get_items` was taking 0.386 sec (i.e. 386 ms), and in this PR, `get_items` is taking only 0.044 sec (i.e. 44 ms) which makes `get_items` 0.342 sec (i.e. 342 ms) faster. And `has_path_to_root` is taking only 0.001 sec (i.e. 1 ms) to execute when we call `get_items` with `include_orphans=False`.