@adampalay 2016-01-04
@Qubad786 , ah, I see.
So the difference in performance on this PR vs. master is 0.001 secs? How big is the course that you're testing on?