@Qubad786 2016-01-04
Yes. It's not that big, It has 2 chapters with 2 subsections each. In each subsection, there are 3 problems, other than that there is 1 orphan problem. But I kept the test course same for both PRs while testing.