@Qubad786 2016-01-05
With the course provided yesterday, master's `get_items` is taking 0.055 sec (i.e. 55 ms) to execute.
  
<img width="839" alt="get_items from master" src="https://cloud.githubusercontent.com/assets/7848408/12117059/e2fe1e6c-b3df-11e5-9f7a-8fb780a60b1d.png">

On this branch `get_items` is taking 0.108 sec (i.e. 108 ms). Single call to `has_path_to_root` is taking 0.015 sec (i.e. 15 ms).

<img width="826" alt="get_items with include_orphans" src="https://cloud.githubusercontent.com/assets/7848408/12117068/f60b8db4-b3df-11e5-83d5-2327276b51e0.png">