@adampalay 2016-01-05
@benpatterson , @ormsbee , @tobz 

This PR gives one the ability to call `get_items` with an optional parameter that filters out orphan modules from its results. For large courses, this introduces a performance penalty, since each module, in order to determine if its parent is in the course, needs to iterate over each module in the course. For smaller courses, this isn't a big deal (the first one @Qubad786 tested on only added 1 millisecond). But for a larger course (the one Hassan tested had around 1500 items), the penalty was steeper. Each check to see that a module wasn't an orphan took 15 ms.

My question for you three is: is this performance hit too high?