@ormsbee 2016-01-05
@adampalay: I guess that really depends on how many things we're calling `get_items` in our worst case scenarios. Is getting rid of a lot of orphan related bugs worth a 50ms worst case? Sure. As long as you can tell me that 50ms is as bad as it gets.

Is there any chance we can just kill the orphans when we're doing the save (so that they never make it into the database in the first place)? Maybe post-processing before we save the structure doc? Or do orphans still need to get exported?