@adampalay 2016-01-05
@ormsbee , I don't know that 50 ms is as bad as it gets — I'd have to see what course has the most amount of items, since this should just be correlated with how many items a course has.

You ask if we can kill orphans we we do a save. Killing orphans turned out to be more complicated than we anticipated (see [this document](https://openedx.atlassian.net/wiki/display/SUST/Fixing+Delete+Orphans)), so we took the tack of just working around them.

Orphans don't get exported, so we have a ticket in to make course imports atomic: [PLAT-863](https://openedx.atlassian.net/browse/PLAT-863). Once that is implemented, you'll be able to remove orphans in a split course by exporting and then importing.