@benpatterson 2016-01-05
I know it doesn't cover the large-course use case you're talking about, but I wanted to capture the bok-choy build results here. The reverted PR added about 70 minutes (of test time...10 mins per shard) to the overall bok-choy run, but this current PR doesn't have an impact.

![image](https://cloud.githubusercontent.com/assets/2004678/12123637/1b3a2152-b3b0-11e5-884a-81ad0b7b0c02.png)

I do think a look at how this impacts the larger courses is worth a cycle.