@Qubad786 2016-01-05
@benpatterson this does not have considerable impact because `get_items` is never used with `include_orphans=False` kwarg anywhere except in only the test which was presented in this PR.