@ormsbee 2016-01-05
@adampalay: Got it. I'm definitely not concerned if it's not being used yet, but if we are switching the LMS over to it, I'd like to see it run on some of the very large MITx courses to gauge impact (e.g. 8.MReVx, 14.74)