@benpatterson 2016-01-05
Yes @Qubad786 I understand that. I just wanted to post additional findings :)