@adampalay 2016-01-06
@ormsbee , for now, we wouldn't switch the whole LMS; it would only be used in a couple of views (grabbing content groups, discussion modules to display on the discussion forum). We'll take a look at the performance impact on MReV too and post results here. So that we're clear, what are the acceptance criteria here?