@ormsbee 2016-01-06
@adampalay: spitballing it, I'd say a ~5% degradation in overall server side request time on those pages is a reasonable tradeoff for getting rid of these types of bugs. If it's > 10%, let's talk about this more.