@Qubad786 2016-01-11
Following are further performance testing with the courses mentioned earlier:

With 8.MReVx; there were total `3405` items in course. 
On master, `get_items` is taking `269 ms` to execute, can be seen in profile below:

<img width="764" alt="master" src="https://cloud.githubusercontent.com/assets/7848408/12232173/b4e02d06-b87e-11e5-94aa-497adc222e00.png">

On this branch, `get_items` is taking `363 ms` from which `has_path_to_root` is taking `106 ms` to execute, can be seen in profile below:

<img width="777" alt="branch" src="https://cloud.githubusercontent.com/assets/7848408/12232212/f0d2790e-b87e-11e5-8964-d80d517f73d0.png">

With 14.74; there were total `2090` items in course. 
On master, `get_items` is taking `122 ms` to execute, can be seen in profile below:

<img width="787" alt="screen shot 2016-01-11 at 4 49 10 pm" src="https://cloud.githubusercontent.com/assets/7848408/12233004/3090b69a-b885-11e5-975a-fdc85e8b6b58.png">

On this branch, `get_items` is taking `159 ms` from which `has_path_to_root` is taking `48 ms` to execute, can be seen in profile below:

<img width="783" alt="branch" src="https://cloud.githubusercontent.com/assets/7848408/12233032/6e1a5f84-b885-11e5-874c-82215ea6ab2d.png">