@adampalay 2016-01-11
@ormsbee , looks like this is a bit more than 10%. As @Qubad786  points out, we're not planning to implement this in that many places, but I also think we can improve the way we calculate which items are in a course tree. (for example, we could cache this list, and regenerate this cache on each publish event). Or we can reimplement `has_path_to_root` where we temporarily store if a module has a path to root so we don't have to walk up the course tree for every module.