@ormsbee 2016-01-13
It doesn't bother me so much that this particular method call goes up in time by that much. When I was spitballing 10%, I meant the overall server execution time for, say, the course index page. Given that, and the limited places where this is going to be used, I'm fine with taking this hit. Thank you for doing this investigative work.